Rename the GetId method of the Model interface to ModelId.
The reason for this is two-fold. First, the Get prefix is not idiomatic
in go. We couldn't just call it Id, because then it would colide with
properties of the same name. Second, zoom uses the same method signature
for its Model interface, making it easy for a struct to satisfy both interfaces.
That's especially helpful for hybrid apps that are using zoom and humble.